1.X/requirements.txt
black==26.1.0
bs4==0.0.2
chromadb==1.5.1
fastapi==0.135.1
google-api-python-client==2.190.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
isort==7.0.0
langchain==1.2.6
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-experimental==0.4.1
langchain-google-community==3.0.5
langchain-mistralai==1.1.1
langchain-tavily==0.2.17
langgraph==1.0.9
langgraph-checkpoint-sqlite==3.0.3
langgraph-supervisor==0.0.31
numpy==2.4.2
mistralai==1.12.4
moviepy==2.2.1
pypdf==6.7.2
PyPDF2==3.0.1
python-dotenv==1.2.2
streamlit==1.53.0
tk==0.1.0
transformers==5.3.0
uvicorn==0.41.0
vt-py==0.22.0